+++ mainfunc/converter.py
@@ -8,7 +8,7 @@ def convert(dir_path, sample_rate, n_electrodes, sep):
     dat_path = dir_path + sep + 'temp_wh.dat'   # kilosort outputs waveforms into temp_wh.dat in the working directory
     # channel_map = np.load(dir_path + sep + 'channel_map.npy')
     spike_clusters = np.load(dir_path + sep + 'spike_clusters.npy')
-    spike_times = np.single(np.load(dir_path + sep + 'spike_times.npy'))
+    spike_times = np.double(np.load(dir_path + sep + 'spike_times.npy'))
     n_clusters = len(np.unique(spike_clusters))
 
     sc = SimpleController(dir_path, dat_path, sample_rate, n_electrodes)
@@ -19,7 +19,7 @@ def convert(dir_path, sample_rate, n_electrodes, sep):
     waveformsL = []
     spike_idsL = []
     primary_electrodeL = []   # map: cluster_id -> channel_id
-    # print('original number of clusters:', n_clusters)    
+    # print('original number of clusters:', n_clusters)
     for cluster_id in range(n_clusters):
         try:
             waveforms, spike_ids = sc._get_waveforms(cluster_id)
